agrobot/requirements.txt
# Core dependencies
google-adk==1.0.0  # Google Agent Development Kit
google-generativeai==0.3.1  # Google Generative AI API
google-ai-generativelanguage==0.6.5  # Google's Generative Language API

# Web server
fastapi==0.105.0  # FastAPI framework for building the API
uvicorn==0.24.0  # ASGI server for FastAPI
websockets==11.0.3  # WebSocket support for real-time communication

# Database
mysql-connector-python==8.3.0  # MySQL connector
sqlglot==19.8.0  # SQL parser and validator

# Utilities
pydantic==2.5.2  # Data validation and settings management
pydantic-settings==2.1.0  # Pydantic settings management
python-dotenv==1.0.0  # Environment variable management
jsonschema==4.19.2  # JSON Schema validation
setuptools==69.0.2  # Package management utilities
typing-extensions==4.8.0  # Type hinting
